refactor(ctimer): encapsulate _inflight, move InflightGuard to ctimer.h
Address Roman's review comments:

1. Make _inflight private with accessors (enterHandler/exitHandler/
   hasInflightHandlers). Restore protected access for _interval, _cstack,
   _signal, _max_timers, _timers (subclass CTimerJvmti needs them).

2. Move InflightGuard into ctimer.h as header-only. Removes engine coupling
   from guards module: guards.cpp no longer #ifdef-includes ctimer.h.
   Non-Linux variant is an empty stub class; the _active bool disappears
   since separate class definitions per platform handle the no-op case.

3. Remove dead Profiler::enableEngines() (replaced by explicit
   _cpu_engine->enableEvents(true) / _wall_engine->enableEvents(true)
   in profiler.cpp). disableEngines() is still used and kept.
