+++ ddprof-lib/src/main/cpp/ctimer.h
@@ -25,23 +25,21 @@
 #include <signal.h>
 
 class CTimer : public Engine {
-protected:
+private:
   // This is accessed from signal handlers, so must be async-signal-safe
   static bool _enabled;
 
-public:
   // Count of signal handlers currently executing past the _enabled check.
-  // stop() drains this to zero before tearing down JFR structures, closing
+  // drainInflight() waits for this to reach zero before JFR teardown, closing
   // the TOCTOU window between the _enabled check and JFR buffer access.
-  // Public so InflightGuard (in guards.cpp) can access it.
   // Placed on its own cache line to avoid false sharing with _enabled:
   // _enabled is read-only on the hot path; _inflight is read-write.
   alignas(64) static int _inflight;
 
-  // Returns true if any handlers are still in-flight (stuck after timeout)
-  static bool hasInflightHandlers() {
-    return __atomic_load_n(&_inflight, __ATOMIC_ACQUIRE) > 0;
-  }
+  // cppcheck-suppress unusedPrivateFunction
+  static void signalHandler(int signo, siginfo_t *siginfo, void *ucontext);
+
+protected:
   static long _interval;
   static CStack _cstack;
   static int _signal;
@@ -51,10 +49,6 @@ public:
 
   int registerThread(int tid);
   void unregisterThread(int tid);
-
-private:
-  // cppcheck-suppress unusedPrivateFunction
-  static void signalHandler(int signo, siginfo_t *siginfo, void *ucontext);
 
 public:
   const char *units() { return "ns"; }
@@ -71,6 +65,19 @@ public:
     __atomic_store_n(&_enabled, enabled, __ATOMIC_RELEASE);
   }
 
+  // In-flight handler tracking accessors (used by InflightGuard).
+  // ACQUIRE on increment / RELEASE on decrement so drainInflight() observes
+  // all handler-side writes before observing the counter at zero.
+  static void enterHandler() {
+    __atomic_fetch_add(&_inflight, 1, __ATOMIC_ACQUIRE);
+  }
+  static void exitHandler() {
+    __atomic_fetch_sub(&_inflight, 1, __ATOMIC_RELEASE);
+  }
+  static bool hasInflightHandlers() {
+    return __atomic_load_n(&_inflight, __ATOMIC_ACQUIRE) > 0;
+  }
+
   // Spin until all signal handlers that passed the _enabled=true check have
   // returned. Must be called with _enabled already false (after disableEngines()),
   // before any JFR teardown that handlers could race against.
@@ -80,6 +87,26 @@ public:
 
   // Get the signal number used by CTimer (0 if not initialized)
   static int getSignal() { return _signal; }
+};
+
+// RAII guard for CTimer signal handler in-flight tracking.
+// Increments the in-flight counter on construction and decrements on destruction,
+// ensuring the counter is always balanced regardless of which exit path the
+// handler takes. The counter (CTimer::_inflight) is cache-line-aligned to avoid
+// false sharing with _enabled, minimizing cache line bouncing.
+//
+// Usage (at the start of CTimer signal handlers):
+//   InflightGuard inflight;
+class InflightGuard {
+public:
+    InflightGuard()  { CTimer::enterHandler(); }
+    ~InflightGuard() { CTimer::exitHandler(); }
+
+    // Non-copyable, non-movable
+    InflightGuard(const InflightGuard&) = delete;
+    InflightGuard& operator=(const InflightGuard&) = delete;
+    InflightGuard(InflightGuard&&) = delete;
+    InflightGuard& operator=(InflightGuard&&) = delete;
 };
 
 // A CPU-time engine that reuses CTimer's per-thread timer_create / SIGPROF
@@ -116,6 +143,16 @@ public:
 
   // No-op on non-Linux platforms
   static bool drainInflight() { return true; }
+  static bool hasInflightHandlers() { return false; }
+};
+
+// No-op InflightGuard on non-Linux platforms where CTimer is unavailable.
+class InflightGuard {
+public:
+    InflightGuard()  {}
+    ~InflightGuard() {}
+    InflightGuard(const InflightGuard&) = delete;
+    InflightGuard& operator=(const InflightGuard&) = delete;
 };
 
 class CTimerJvmti : public Engine {
