+++ ddprof-lib/src/main/cpp/profiler.h
@@ -126,7 +126,6 @@ private:
   void switchLibraryTrap(bool enable);
   static void prewarmUnwinder();
 
-  void enableEngines();
   void disableEngines();
 
   void onThreadStart(jvmtiEnv *jvmti, JNIEnv *jni, jthread thread);
